add checking for id (isPresent(long id)) and change delete method from GET to POST

+++ src/main/java/recordstore/controller/AdminArtistController.java
@@ -6,7 +6,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import recordstore.DTO.ArtistDTO;
 import recordstore.entity.Artist;
+import recordstore.mapper.ArtistMapper;
 import recordstore.service.ArtistService;
 
 import javax.servlet.http.HttpServletRequest;
@@ -21,9 +23,12 @@ import java.util.stream.IntStream;
 @RequestMapping("/admin/artists/")
 public class AdminArtistController{
 
+    private final ArtistMapper artistMapper;
+
     private final ArtistService service;
 
-    public AdminArtistController(ArtistService service) {
+    public AdminArtistController(ArtistMapper artistMapper, ArtistService service) {
+        this.artistMapper = artistMapper;
         this.service = service;
     }
 
@@ -38,38 +43,44 @@ public class AdminArtistController{
 
     @GetMapping("/add")
     public String showAddForm(Model model){
-        model.addAttribute("newArtist", new Artist());
+        model.addAttribute("newArtist", new ArtistDTO());
         return "admin/artists/add";
     }
 
     @PostMapping("/add")
-    public String saveArtist(@Valid @ModelAttribute("newArtist") Artist artist, BindingResult result) throws IOException {
+    public String saveArtist(@Valid @ModelAttribute("newArtist") ArtistDTO artistDTO, BindingResult result) throws IOException {
         if (result.hasErrors()) {
             return "admin/artists/add";
         }
-        service.saveArtist(artist);
+        service.saveArtist(artistMapper.fromDTO(artistDTO));
         return "redirect:/admin/artists/";
     }
 
     @GetMapping("/edit/{id}")
     public String showEditForm(@PathVariable long id, Model model) {
-        Artist artist = service.getArtist(id);
-        model.addAttribute("artist", artist);
+        if (service.isPresent(id)) {
+            ArtistDTO artistDTO = artistMapper.toDTO(service.getArtist(id));
+            model.addAttribute("artist", artistDTO);
+        } else {
+            model.addAttribute("error", "Artist is not found");
+        }
         return "admin/artists/edit";
     }
 
     @PostMapping("/edit/{id}")
-    public String updateArtist(@Valid @ModelAttribute("artist") Artist artist, BindingResult result) throws IOException {
+    public String updateArtist(@Valid @ModelAttribute("artist") ArtistDTO artistDTO, BindingResult result) throws IOException {
         if (result.hasErrors()) {
             return "admin/artists/edit";
         }
-        service.saveArtist(artist);
+        service.saveArtist(artistMapper.fromDTO(artistDTO));
         return "redirect:/admin/artists/";
     }
 
-    @GetMapping("delete/{id}")
-    public String delete(@PathVariable long id) throws IOException {
-        service.deleteArtist(id);
+    @PostMapping("/delete")
+    public String delete(@RequestParam("id") long id) throws IOException {
+        if (service.isPresent(id)){
+            service.deleteArtist(id);
+        }
         return "redirect:/admin/artists/";
     }
